Check for wrong changelog changes
